Moved TypedSeq and Respondent to `utils`
- Created utils folder for the general utilities I created while implementing Dote, i.e. TypedSeq and Respondent.
- Rake task added for running tests on utils files only. I didn't realise that these tests weren't included in the default rake task so they were requiring the wrong file paths for a long while.
