+++ utils/typed_seq.rb
